upto kue

+++ controllers/posts_controller.js
@@ -3,10 +3,20 @@ const Comment = require('../models/comment')
 
 module.exports.create = async function (req, res) {
     try {
-        await Post.create({
+        let post = await Post.create({
             content: req.body.content,
             user: req.user._id
         })
+
+        if (req.xhr) {
+            return res.status(200).json({
+                data: {
+                    post: post
+                },
+                message: "Post created!"
+            });
+        }
+
         req.flash('success', 'You Added the post')
         return res.redirect('back')
     }
@@ -20,6 +30,16 @@ module.exports.destroy = async function (req, res) {
         let post = await Post.findById(req.params.id);
         if (post.user == req.user.id) {
             post.deleteOne();
+
+            if (req.xhr) {
+                return res.status(200).json({
+                    data: {
+                        post_id: req.params.id
+                    },
+                    message: 'post deleted'
+                })
+            }
+
             req.flash('success', 'You delete the post')
             await Comment.deleteMany({ post: req.params.id });
             return res.redirect('back')
